feat: show friendly display names for assets
Assets created by the AWS Construct Library now use friendly display names.

- Assets like stack templates now show a display name like `MyStack Template`.
- Assets from `Asset` or `ImageAsset` constructs now show the
  stack-relative construct path of the asset construct, like
  `MyConstruct/MyFunction/Code`.

These defaults explicitly don't use entire construct paths on purpose,
because those will be confusing in combination with the deduplication
behavior of assets; if the same asset is added in multiple places in the
construct tree, it will be deduplicated (based on the asset hash) and
only processed once. The defaults have therefore been chosen to give
useful contextual information, without implying a context that will
alarm users.

For example, it would be confusing for users to see references to
`/MyBetaStage/MyStack Template` or `/MyBetaStack/MyFunction/Code` when
they were deploying a "prod" environment; but if we used full context
paths and the Beta resources were defined first, that would be the
default behavior.

A `displayName` property is available to override the default if the
defaults are undesirable. This property is available for
Asset/AssetImage constructs and their helper functions, not for template
assets.