fix: context provider's `ignoreErrorOnMissingContext` parameter is misleading
In `ContextProvider.getValue()`, `ignoreErrorOnMissingContext` is a
request to the CLI's context provider to not fail the lookup, but return
the dummy value instead.

The operation doesn't have anything to do with missing context, and
missing context isn't an error. Deprecate that parameter and add
`ignoreFailedLookup` instead.

Also explain the `dummyValue` field and this one a bit better.