chore(eks): update default capacity type from NODEGROUP to AUTOMODE in the doc (#33886)
### Issue # (if applicable)

Closes https://github.com/aws/aws-cdk/issues/33884

### Reason for this change

The reason for this change is to correct the default value of the defaultCapacityType property in the AWS CDK's EKS V2 Alpha package. The current documentation incorrectly states that the default value is NODEGROUP, while the actual default value used in the code is AUTOMODE. This can cause confusion and errors for users.


### Description of changes

Update the JSDoc comment in the cluster.ts file to correctly reflect that the default value of the `defaultCapacityType` property is `AUTOMODE`.


### Describe any new or updated permissions being added




### Description of how you validated changes



### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*