fix: context provider's `ignoreErrorOnMissingContext` parameter is misleading (#33875)
In `ContextProvider.getValue()`, `ignoreErrorOnMissingContext` is a request to the CLI's context provider to not fail the lookup, but return the dummy value instead.

The operation doesn't have anything to do with missing context, and missing context isn't an error. Deprecate that parameter and add `mustExist` instead (with reversed semantics).

Also explain the `dummyValue` field and this one a bit better.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*