chore: add addedon field automation in prioritization board (#33856)
### Issue # (if applicable)

NA

### Reason for this change

Currently in the Prioritization dasboard, we don't have a way to filter the PRs and Issues by date when it is added.

### Description of changes

This PR enables automation for adding current date to the `AddedOn` column field in the Github Prioritization dasboard whenever new PR/Issues are being added to the dashboard.

### Describe any new or updated permissions being added

NA


### Description of how you validated changes

Tested in test dashboard.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*