feat: show friendly display names for assets (#33844)
Assets created by the AWS Construct Library now use friendly display names.

- Assets like stack templates now show a display name like `MyStack Template`.
- Assets from `Asset` or `ImageAsset` constructs now show the stack-relative construct path of the asset construct, like `MyConstruct/MyFunction/Code`.

These defaults explicitly don't use entire construct paths on purpose, because those will be confusing in combination with the deduplication behavior of assets; if the same asset is added in multiple places in the construct tree, it will be deduplicated (based on the asset hash) and only processed once. The defaults have therefore been chosen to give useful contextual information, without implying a context that will alarm users.

For example, it would be confusing for users to see references to `/MyBetaStage/MyStack Template` or `/MyBetaStack/MyFunction/Code` when they were deploying a "prod" environment; but if we used full context paths and the Beta resources were defined first, that would be the default behavior.

A `displayName` property is available to override the default if the defaults are undesirable. This property is available for Asset/AssetImage constructs and their helper functions, not for template assets.

## Implementation notes

Since in the cloud assembly `displayName` is part of neither asset *source* not asset *destination*, but an additional field, I've had to add a 3rd parameter `Add*AssetOptions` to most of the `add*Asset()` methods and it needs to be forwarded in a number of places.

The CDK Pipelines implementation also needed to add the concept of a `displayName` to pipeline graph nodes, since reflecting it in the node IDs had far-reaching consequences (the node ID also affects the construct IDs of constructs that get created, and we don't want to have to recreate these just for a different display name). Uniquification of action names happens at the very last moment, when we put the Pipeline together.

`GraphNode.of()` accepts positional parameters so I added a positional optional argument for display name there, but since in TypeScript subclasses have to be structurally compatible with superclasses (😶‍🌫️) and `Graph.of()` already takes an argument in that position, the type of that argument is now weird. The alternative would be to change the signature to `Graph.of(id, data, <displayName?>, nodes?)`, inserting `displayName`, but even though this is not part of the super public jsii interface this method is still accessible on the semi-public "extend me via TypeScript only" interface and we don't want to unnecessarily break consumers.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*