fix(core): implicit Aspect applications do not override custom Aspect applications
Some CDK methods apply mutatinhg Aspects on behalf of users. Since
https://github.com/aws/aws-cdk/pull/32333, these Aspects have a priority
of `MUTATING` to classify their behavior.

If a user-applied Aspect (priority `DEFAULT`) now configures the same
property as an implicitly added Aspect:

* Before that change, the relative execution order depended on the
  location of the Aspects in the construct tree.
* After that change, the user Aspect always "wins" (executes last)
  because its priority is higher.

In this change, we introduce a feature flag which gives the Aspects
introduced in https://github.com/aws/aws-cdk/pull/32333 that users are
likely to override a priority only if the feature flag is enabled. This
change does not affect users have given their own Aspects a `MUTATING`
priority as well since 2.172.0.

Did not touch the following Aspects:

- In `integ-tests-alpha`: overriding logical IDs in assertions stacks
  does not affect production infrastructure.
- Tags: tags are exclusively manipulated through the official APIs,
  so there no conflict between custom and implicit Aspects.
- CDK Pipelines: there cannot be a conflict because the customer
  can't create a default pipeline before the implicit Aspect.