feat(ec2): adding `placementGroup` to `LaunchTemplateProps` and `LaunchTemplate` (#33726)
### Issue # (if applicable)

Closes #33721

### Reason for this change

To allow for passing `PlacementGroup`s to `LaunchTemplate`s


### Description of changes

Adding `placementGroup` to `LaunchTemplateProps`. I chose to only implement this part of the `placement` field in LaunchTemplates, not doing anything with dedicated hosts or the like, as those don't have any CDK constructs. This change enables creating a `PlacementGroup` and passing it to a `LaunchTemplate`

### Describe any new or updated permissions being added

None


### Description of how you validated changes

Added a unit test, ran the existing tests as well.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*