fix(core): asset names for nested stacks contain Tokens
Since Nested Stack names are Tokens (resolving to `{ Ref:
'AWS::StackName' }`), if we use `nestedStack.stackName` as the asset
name just like we do for regular stacks, we end up with Tokens in the
display name of the nested stack template asset.

This is pointless because the token will not be resolved and look like
`${Token[TOKEN.639]}`; plus, it breaks tests because the Token numbers
are different on every run. In CDK Pipelines, it would also lead to the
pipeline restarting on every pipeline run.

Instead, test for and reject Tokens, and use the nested stack's
construct path instead.