fix: customer aspect cannot add Tags if a BucketNotifications construct is present (#33979)
The title refers to a specific case in which we discovered this, but the problem was more general.

Now that Aspects have a priority, we want to execute based on priority as the primary sort key. To make sure we don't make mistakes, the new AspectsV2 invocation (enabled with flag `@aws-cdk/core.aspectStabilization: true`) throws an error if it detects a violation of the strict priority execution order.

Because of the moment aspects are looked up and sorted by priority, this currently goes wrong if:

- Aspects with a low and high priority order (say, 100 and 700) are already present in the construct tree
- The invocation of the low-priority Aspect adds another Aspect with a priority in between 100 and 700.

The reason is that we used to look up the aspects before iterating, and iterate over them. At that point we have the list `[100, 700]`, and we invoke both. Then, when we go around again, the new list is `[100, 500, 700]` with the `500` Aspect having not been invoked yet -- but also the Aspect with priority `700` already having been invoked. If we invoke `500` at that point, we would have inverted the priority order, so we fail instead.

That is undesirable: what we *should* have done is noticed that `500` got inserted between `100` and `700` and do that next.

Currently achieve that with the following:

- We associate a global revision number with the entire construct tree, at the root.
- Every time an aspect gets added, we increment that number.
- This allows us to detect Aspect order cache validations, so we can abort what we're doing and recalculate the tree.

The system of inserting and recalculating could probably all be made more efficient, but aspects adding aspects should be relatively rare so until we have evidence that it is too expensive, this will do for now.

In the specific case reported for this, a customer Aspect with priority 10 adds a Tag Aspect (priority 200) while the `BucketNotification` construct also adds an Aspect on itself with priority 500. The mere presence of this `BucketNotification` construct with its self-added Aspect prevents any Aspect from adding tags.

Closes #33943.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*