fix(pipelines): can't have the same asset display name 3 times
When introducing friendly asset names in #33844, we made a mistake in
calculating unique action names for distinct assets with the same
display name.

As a result, the same asset could be added to the pipeline twice,
but not thrice. Fix the logic.

Also in this PR: Stages can't have a name of only one character. Fix
that.

Closes #34004.