Change the json processing strategy.
We are now using a safe json value for marker replacement and avoid loading the complete JSON file in memory.

This change also add a way to trigger the new processing when creating the source data. This now limit the processing at the asset level instead of the deployment level.