fix(ec2-alpha): update default config for Subnet's `assignIpv6AddressOnCreation` (#34116)
### Issue # (if applicable)

N/A

### Reason for this change

The `Subnet.assignIpv6AddressOnCreation` attribute was being set in the L1, despite it not being necessary to set. This may cause issues with backwards migrations from V1 subnets to V2, since there will be this attribute difference across both.

### Description of changes

```ts
const subnet = new CfnSubnet(this, 'Subnet', {
  ...
  assignIpv6AddressOnCreation: props.assignIpv6AddressOnCreation /* ?? false */,
  mapPublicIpOnLaunch: props.mapPublicIpOnLaunch /* ?? undefined */ // this one was just a redundant nit
}
```

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

`yarn test && yarn integ-runner --update-on-failed --dry-run` (dry run since it's a very minor change, and some of the integ tests for IPv6 and IPAM are finnicky)

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*