chore(cognito-identitypool): make role attachment resource public (#34234)
### Issue # (if applicable)

N/A

### Reason for this change

In order to allow for more flexibility when configuring Identity Pool resources, one may need to override attributes in the role attachment. To make this process easier, we should be able to access this resource the same way as we would access the L2 resource.

### Description of changes

```ts
  // private readonly roleAttachment: IdentityPoolRoleAttachment;
  public readonly roleAttachment: IdentityPoolRoleAttachment;
```

### Describe any new or updated permissions being added

N/A

### Description of how you validated changes

`yarn build && yarn test aws-cognito-identitypool`

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*