chore: make all L2 Constructs property injectable during release (#34328)
### Issue # (if applicable)

This is a followup to https://github.com/aws/aws-cdk/pull/33887.

DO NOT merge this PR until after 33887 is merged.

### Reason for this change

In the original Property Injection PR, we enabled 122 Constructs for Property Injections, but we want to make this feature available for all L2 Constructs.

### Description of changes

Add logic to Constructs ConstructsUpdater to if a Construct already have PROPERTY_INJECTION_ID property.  If not, it will:
- Add PROPERTY_INJECTION_ID with value that is calculated from path and class name
- Import aws-cdk-lib/core/lib/prop-injectable
- Add class decorator @propertyInjectable

### Describe any new or updated permissions being added

No permission change.

### Description of how you validated changes

- Added unit tests and ran `npm test`
- Ran `./bin/update-construct-metadata` locally and manually verified a few random files.

### Checklist
- [ ] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*